Rename util/Date.ts getDateFormat -> getMomentDateFormat and getDateTimeFormat -> getMomentDateTimeFormat
