+++ packages/components/src/internal/app/utils.ts
@@ -45,7 +45,6 @@ import {
     WORKFLOW_HOME_HREF,
     WORKFLOW_KEY,
 } from './constants';
-import { Container } from '../components/base/models/Container';
 
 // Type definition not provided for event codes so here we provide our own
 // Source: https://www.iana.org/assignments/websocket/websocket.xml#close-code-number
@@ -564,16 +563,6 @@ function getApplicationUrlBase(moduleName: string, currentProductId: string, mod
     return !lcProductId || lcProductId === currentProductId.toLowerCase()
         ? ''
         : buildURL(lcProductId, 'app.view', undefined, { returnUrl: false });
-}
-
-// CONSIDER: Moving this to util/Date.ts after rectifying name collision with similar method already declared there.
-export function getDateFormat(container?: Partial<Container>): string {
-    return (container ?? getServerContext().container).formats.dateFormat;
-}
-
-// CONSIDER: Moving this to util/Date.ts after rectifying name collision with similar method already declared there.
-export function getDateTimeFormat(container?: Partial<Container>): string {
-    return (container ?? getServerContext().container).formats.dateTimeFormat;
 }
 
 // Returns the friendly name of the product, primarily for use in help text.
